changed a ton of things
switched control and command home row mods, I use control pretty often with vim and command less, so gonna give my pinkies a rest. the command on the pinky also works better for some one handed shortcuts, like hitting command + number to go to a browser tab with the right hand.
num layer is more normal again, and moved to the right hand, with the left hand a near dup of the sym layer, with added comma and dot taking the place of amps and pipe, which aren't used often when typing out numerical expressions
sym layer positioning tweaked a bit on the left hand to make bigrams easier, added autopair macros to the right hand. right hand is set up to overwrite less from the base layer and to play with the num layer better where numpad doesn't overlap
got rid of the question mark to bang morph, I use the slash often enough that I want it on the base layer, and the bang is still easy to get to on the sym layer
fn layer also moved to match the num layer
added clipboard and undo/redo commands to the nav and mouse layers
got rid of the hold functionality on arrow keys in the nav layer
changed all left hand home keys in the nav layer to sticky mods, changed mouse layer to plain mods as well
quick layer is gone, with some of its functionality moved into the left hand of the mouse layer
mod/bt and media layers have been combined into one layer, with hold keys for bluetooth profile selection on the bottom row of left hand